NNUE integrated with engine
now can evaluate moves and pick the "best one"

without alpha-beta or some other search it's still bad

+++ src/nnue/nnue.h
@@ -33,7 +33,7 @@ public:
 	NNUE();
 	void get_active_features(const Board b);
 	void refresh_accumulator();
-	float eval();
+	int eval();
 
 };
